Vertical scrollbars and Fl_Numeric_Input did mousewheel movement backwards.
Changed the arguments for drawing and time from double to float. This
seems to be more popular for drawing libraries and is faster. Big
problem is that C++ will not resolve foo(int,int) and foo(float,float)
unless you identify both arguments as floats or both are
integers. Also VC++ warns about the int->float conversions. Both of
these can be fixed with casts but I have not gotten all the VC++
warnings out yet (I don't turn them off because the double->float
warning is also turned off and I want that one).


git-svn-id: file:///fltk/svn/trunk@2623 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/Enumerations.h
@@ -1,5 +1,5 @@
 //
-// "$Id: Enumerations.h,v 1.8 2002/09/02 06:33:46 spitzak Exp $"
+// "$Id: Enumerations.h,v 1.9 2002/09/16 00:29:05 spitzak Exp $"
 //
 // Enumerations for the Fast Light Tool Kit (FLTK).
 //
@@ -33,7 +33,7 @@
 // because the old "const double" definition would not allow for conditional
 // compilation...
 //
-// FL_VERSION is a double that describes the major, minor, and patch version
+// FL_VERSION is a float that describes the major, minor, and patch version
 // numbers.  Version 1.1 is actually stored as 1.01 to allow for more than 9
 // minor releases.  Similarly, version 1.0.5 is stored as 1.0005.
 //
@@ -45,9 +45,9 @@
 #define FL_MAJOR_VERSION	2
 #define FL_MINOR_VERSION	0
 #define FL_PATCH_VERSION	1
-#define FL_VERSION		((double)FL_MAJOR_VERSION + \
-				 (double)FL_MINOR_VERSION * 0.01 + \
-				 (double)FL_PATCH_VERSION * 0.0001)
+#define FL_VERSION		(FL_MAJOR_VERSION + \
+				 FL_MINOR_VERSION * 0.01f + \
+				 FL_PATCH_VERSION * 0.0001f)
 
 typedef unsigned char uchar;
 typedef unsigned long ulong;
@@ -232,5 +232,5 @@ enum {
 #endif
 
 //
-// End of "$Id: Enumerations.h,v 1.8 2002/09/02 06:33:46 spitzak Exp $".
+// End of "$Id: Enumerations.h,v 1.9 2002/09/16 00:29:05 spitzak Exp $".
 //
